requirements.txt
djangorestframework
django
django-crispy-forms
